init renderer module
