--- __mocks__/template.js
@@ -1,1 +1,0 @@
-module.exports = function(e){return'<div id="15c8ed8a-d4d6-49e4-9f3a-fb612e89679f"></div><script>var targetNode = document.getElementById("15c8ed8a-d4d6-49e4-9f3a-fb612e89679f");targetNode.setAttribute("id","");var reactApp = (function(a){function b(d){if(c[d])return c[d].exports;var f=c[d]={i:d,l:!1,exports:{}};return a[d].call(f.exports,f,f.exports,b),f.l=!0,f.exports}var c={};return b.m=a,b.c=c,b.i=function(d){return d},b.d=function(d,f,g){b.o(d,f)||Object.defineProperty(d,f,{configurable:!1,enumerable:!0,get:g})},b.n=function(d){var f=d&&d.__esModule?function(){return d["default"]}:function(){return d};return b.d(f,"a",f),f},b.o=function(d,f){return Object.prototype.hasOwnProperty.call(d,f)},b.p="",b(b.s=2)})([function(a,b,c){"use strict";function f(n,o){if(!(n instanceof o))throw new TypeError("Cannot call a class as a function")}function g(n,o){if(!n)throw new ReferenceError("this hasn\\"t been initialised - super() hasn\\"t been called");return o&&("object"==typeof o||"function"==typeof o)?o:n}function h(n,o){if("function"!=typeof o&&null!==o)throw new TypeError("Super expression must either be null or a function, not "+typeof o);n.prototype=Object.create(o&&o.prototype,{constructor:{value:n,enumerable:!1,writable:!0,configurable:!0}}),o&&(Object.setPrototypeOf?Object.setPrototypeOf(n,o):n.__proto__=o)}Object.defineProperty(b,"__esModule",{value:!0});var j=function(){function n(o,p){for(var r,q=0;q<p.length;q++)r=p[q],r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(o,r.key,r)}return function(o,p,q){return p&&n(o.prototype,p),q&&n(o,q),o}}(),k=c(1),l=function(n){return n&&n.__esModule?n:{default:n}}(k),m=function(n){function o(){f(this,o);var q=g(this,(o.__proto__||Object.getPrototypeOf(o)).call(this));return q.state={count:0},q}return h(o,n),j(o,[{key:"onClick",value:function(){this.setState({count:this.state.count+1})}},{key:"render",value:function(){return l.default.createElement("div",null,l.default.createElement("h1",null,this.state.count),l.default.createElement("button",{onClick:this.onClick.bind(this)},"Count up ",this.props.name,"!!"))}}]),o}(l.default.Component);b.default=m},function(a){a.exports=React},function(a,b,c){"use strict";var f=c(0),g=function(h){return h&&h.__esModule?h:{default:h}}(f);a.exports=function(h){return React.createElement(g.default,h,null)}}]);;ReactDOM.render(reactApp('+JSON.stringify(e)+"), targetNode);</script>"};
