--- blueprint/src/README.md
@@ -1,2 +1,0 @@
-# oc-template-react base component
-
