--- blueprint/src/app.js
@@ -1,3 +1,0 @@
-import React from 'react';
-const app = ({ name }) => <h1>Hello {name}!</h1>
-export default app;
