--- blueprint/src/server.js
@@ -1,4 +1,0 @@
-export const data = (context, callback) => {
-  const { name }  = context.params;
-  return callback(null, { name });
-};
