+++ packages/oc-template-react/README.md
@@ -1,0 +1,3 @@
+# oc-template-react - Renderer module
+
+OC template to handle components of type `oc-template-react`
