+++ packages/oc-template-react/__tests__/getInfo.js
@@ -1,0 +1,7 @@
+const getInfo = require("../lib/getInfo");
+
+test("should return the correct info", () => {
+  const info = getInfo();
+  delete info.version;
+  expect(info).toMatchSnapshot();
+});
