+++ packages/oc-template-react/index.js
@@ -1,0 +1,11 @@
+"use strict";
+
+const getCompiledTemplate = require("./lib/getCompiledTemplate");
+const getInfo = require("./lib/getInfo");
+const render = require("./lib/render");
+
+module.exports = {
+  getInfo,
+  getCompiledTemplate,
+  render
+};
