+++ packages/oc-template-react/lib/getInfo.js
@@ -1,0 +1,10 @@
+const packageJson = require("../package.json");
+const externals = Object.keys(packageJson.externals).map(
+  dep => packageJson.externals[dep]
+);
+
+module.exports = () => ({
+  type: packageJson.name,
+  version: packageJson.version,
+  externals
+});
