+++ packages/oc-template-react/lib/render.js
@@ -1,0 +1,8 @@
+module.exports = (options, callback) => {
+  try {
+    const html = options.template(options.model);
+    return callback(null, html);
+  } catch (error) {
+    return callback(error);
+  }
+};
