--- src/index.js
@@ -1,11 +1,0 @@
-const getInfo = require('./getInfo');
-const getCompiledTemplate = require('./getCompiledTemplate');
-const compile = require('./compile');
-const render = require('./render');
-
-module.exports = {
-  getInfo,
-  getCompiledTemplate,
-  compile,
-  render
-};
