--- src/render.js
@@ -1,2 +1,0 @@
-module.exports = (options, callback) =>
-  callback(null, options.template(options.model));
